requirements.txt
comtypes==1.4.11
configparser==7.2.0
MouseInfo==0.1.3
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
psutil==7.0.0
PyAutoGUI==0.9.54
pycaw==20240210
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
pytweening==1.2.0
